+++ resources/views/page/create.blade.php
@@ -12,9 +12,8 @@
                 value="{{ $context->valueOf('title') }}" width2="4" />
             <x-laraknife.forms.string position="last" name="name" label="Name (URL)"
                 value="{{ $context->valueOf('name') }}" width2="4" />
-                <x-laraknife.forms.combobox position="first" name="language_scope" label="Language" :options="$optionsLanguage" width2="4" />
-                <x-laraknife.forms.string position="last" name="column" label="Columns"
-                value="{{ $context->valueOf('columns') }}" width2="4" />
+            <x-laraknife.forms.combobox position="alone" name="language_scope" label="Language" :options="$optionsLanguage"
+                width2="4" />
             <x-laraknife.forms.text position="alone" name="contents" label="Contents"
                 value="{{ $context->valueOf('contents') }}" width2="10" rows="10" />
             <x-laraknife.forms.text position="alone" name="info" label="Info" value="{{ $context->valueOf('info') }}"